pyproject.toml
[project]
name = "hackathon-gazprom-2026"
version = "0.1.0"
description = "PDF → Markdown pipeline: DocLayout-YOLO + Docling (векторные таблицы) + DeepSeek-OCR-2 (растр)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Layout detection
    "doclayout-yolo>=0.0.4",
    "ultralytics>=8.4.38",
    "torch==2.6.0",
    "torchvision==0.21.0",
    "transformers>=4.50.0",
    "tokenizers>=0.21.0",
    "accelerate>=1.0.0",
    "einops",
    "addict",
    "easydict",
    "docling>=2.0.0",
    "docling-core>=2.0.0",
    "pymupdf>=1.24.0",
    "pillow>=10.0.0",
    "opencv-python>=4.9.0",
    "numpy>=1.26.0",
    "huggingface-hub>=0.24.0",
]
